Update development with hot fix and changelog (#44)
* Add CL items for 1.3 release

* try to find existing ms sample in case of replicate data

* Add CL information for patch 1.3.1

Affects the register-mzml-dropbox. Checks the
given project for existing samples with the provided
sample code of type "Q_MS_RUN". A new sample will be only
created in openBIS, if this sample does not yet exist.

* Provide format documentation for single-end / paired-end data registration (#40)

This CL introduces some documentation for the data structure in preparation to transfer data to QBIC
and register them into qPortal.

* Update Changelog

Co-authored-by: Sven Fillinger <sven.fillinger@qbic.uni-tuebingen.de>
Co-authored-by: Sven F <sven.fillinger@mailbox.org>